Added option to disable Windows Recall AI snapshots (#61) Removed code for migrating LastSettings file to SavedSettings

+++ Win11Debloat.ps1
@@ -23,6 +23,7 @@ param (
     [switch]$HideSearchTb, [switch]$ShowSearchIconTb, [switch]$ShowSearchLabelTb, [switch]$ShowSearchBoxTb,
     [switch]$HideTaskview,
     [switch]$DisableCopilot,
+    [switch]$DisableRecall,
     [switch]$DisableWidgets,
     [switch]$HideWidgets,
     [switch]$DisableChat,
@@ -544,22 +545,9 @@ foreach ($Param in $SPParams) {
     }
 }
 
-# Check if SavedSettings file exists, if it doesn't exist check if LastSettings file exists
-if (Test-Path "$PSScriptRoot/SavedSettings") {
-    if ([String]::IsNullOrWhiteSpace((Get-content "$PSScriptRoot/SavedSettings"))) {
-        # Remove SavedSettings file if it's empty
-        Remove-Item -Path "$PSScriptRoot/SavedSettings" -recurse
-    }
-}
-elseif (Test-Path "$PSScriptRoot/LastSettings") {
-    if ([String]::IsNullOrWhiteSpace((Get-content "$PSScriptRoot/LastSettings"))) {
-        # Remove LastSettings file if it's empty
-        Remove-Item -Path "$PSScriptRoot/LastSettings" -recurse
-    }
-    else {
-        # Rename LastSettings file to SavedSettings if it isn't empty
-        Rename-Item -Path "$PSScriptRoot/LastSettings" -NewName "$PSScriptRoot/SavedSettings"
-    }
+# Remove SavedSettings file if it exists and is empty
+if (Test-Path "$PSScriptRoot/SavedSettings" -and [String]::IsNullOrWhiteSpace((Get-content "$PSScriptRoot/SavedSettings"))) {
+    Remove-Item -Path "$PSScriptRoot/SavedSettings" -recurse
 }
 
 # Only run the app selection form if the 'RunAppConfigurator' parameter was passed to the script
@@ -751,6 +739,12 @@ if ((-not $global:Params.Count) -or $RunDefaults -or $RunWin11Defaults -or ($SPP
 
                 if ($( Read-Host -Prompt "Disable Windows Copilot? This applies to all users (y/n)" ) -eq 'y') {
                     AddParameter 'DisableCopilot' 'Disable Windows copilot'
+                }
+
+                Write-Output ""
+
+                if ($( Read-Host -Prompt "Disable Windows Recall snapshots? This applies to all users (y/n)" ) -eq 'y') {
+                    AddParameter 'DisableRecall' 'Disable Windows Recall snapshots'
                 }
             }
 
@@ -1129,6 +1123,10 @@ else {
         }
         'DisableCopilot' {
             RegImport "> Disabling Windows copilot..." $PSScriptRoot\Regfiles\Disable_Copilot.reg
+            continue
+        }
+        'DisableRecall' {
+            RegImport "> Disabling Windows Recall snapshots..." $PSScriptRoot\Regfiles\Disable_AI_Recall.reg
             continue
         }
         {$_ -in "HideWidgets", "DisableWidgets"} {
